Dodao komentare i zavrsio bota za licitiranje

+++ Faza5/Implementacija/app/Http/Controllers/dimitrije/AdminController.php
@@ -1,5 +1,5 @@
 <?php
-
+//Dimitrije Plavsic 18/220
 namespace App\Http\Controllers\dimitrije;
 
 use App\Http\Controllers\Controller;
@@ -8,26 +8,37 @@ use App\Models\dimitrije\RegistredModel;
 use App\Models\dimitrije\ModeratorModel;
 use App\Models\dimitrije\AdministratorModel;
 use App\Http\Controllers\dimitrije\ModeratorController;
+use App\Models\nikola\Auction;
+use App\Models\nikola\Exhibition;
 
+/* klasa za funkcionalnosti administratora */
 
 class AdminController extends Controller
 {
+    //kreiranje nove instance i dodavanje middleware-a
     function __construct() {
         $this->middleware('admin', ['except' => [
             'testing'
         ]]);
     }
 
+    //metoda koja izbacuje view za brisanje naloga od strane admina
     public function adminDeleteAccount()
     {
         return view('dimitrije/adminDeleteAccount', ['body_id' => 'aboutus_body']);
     }
+
+    //submit unetog korisnika kojeg admin zeli obrisati
     public function adminDeleteAccountSubmit(Request $request)
     {
         $username = $request['usernameInput'];
         $users = RegistredModel::all();
         foreach ($users as $user) {
             if($user['Username'] == $username) {
+                if(Auction::isUserBidder($user['IDUser'])) return back()->with('status', "You cannot delete an account while they're the highest bidder at an auction!");
+                else if ( Auction::isUserOwner($user['IDUser'])) return back()->with('status', "You cannot delete an account while they're the owner of an auction!");
+                else if(Exhibition::isUserOwner($user['IDUser'])) return back()->with('status', "You cannot delete an account while they're the owner of an exhibition!");
+
                 $isAdmin = AdministratorModel::find($user['IDUser']);
                 if($isAdmin == null){
                     $user->delete();
@@ -43,12 +54,13 @@ class AdminController extends Controller
         return back()->with('status', $ret);               
     }
 
+    //metoda koja izbacuje view za ukidanja uloga moderatora od strane admina
     public function downgradeModerator()
     {
         return view('dimitrije/downgradeModerator');
     }
 
-
+    //submit unetog moderatora kojem admin zeli ukinuti uloge
     public function downgradeModeratorSubmit(Request $request){
 
         $users = RegistredModel::all();
@@ -77,6 +89,7 @@ class AdminController extends Controller
         return back()->with('status', $ret);
     }
 
+    //metoda za testiranje sesija
     public function testing(){
         return dd(session()->getDrivers()['file']);
     }
